Move test to be make instead of ui
