--- tests/ui/crate-loading/auxiliary/dep-2-reexport.rs
@@ -1,4 +1,0 @@
-//@ edition:2021
-//@ aux-build:multiple-dep-versions-2.rs
-extern crate dependency;
-pub use dependency::do_something;
