--- tests/ui/crate-loading/auxiliary/multiple-dep-versions-1.rs
@@ -1,7 +1,0 @@
-#![crate_name="dependency"]
-//@ edition:2021
-//@ compile-flags: -C metadata=1 -C extra-filename=-1
-pub struct Type;
-pub trait Trait {}
-impl Trait for Type {}
-pub fn do_something<X: Trait>(_: X) { }
